requirements.txt
tensorflow>=2.10.0
tensorflow-keras
sentence-transformers>=2.0.0
openai>=1.0.0
lxml>=4.6.0
androguard>=3.3.5
numpy>=1.21.0
matplotlib>=3.5.0
requests>=2.28.0
python-dotenv>=0.19.0
pandas>=1.3.0
scikit-learn>=1.0.0
seaborn>=0.11.0